break up cli/log/mod.rs
